+++ CHANGELOG.md
@@ -8,6 +8,7 @@ A preview of major changes can be found in the Wiki ([Latest Changes](https://gi
 - Display visits in reverse order on spectate page
 - Additional tournament statistics displayed on statistics page
 - New view for `tournaments/<id>/obs` which shows a counter for certain statistics in a given tournament
+- More vocal announcements
 
 #### Changed
 - Added JDC shirt to spectate page